\name{Internal_data}
\alias{ctn.plot.data}
\alias{mix.plot.data}
\alias{mse.data}
\alias{pred.mix.rs}
\alias{pred.prs.std}
\alias{y.test.bin}
\alias{y.test.mix}
\title{Internal Data Object}
\description{
The data is for internal use, and is not meant for users.
}
\details{
For speeding up vignette build purpose.
}
\format{
  Data objects used for demonstration
}
\keyword{datasets}
